refactor: create `vram` module
